made optimization: removed unused views and set to null on ondestroyview and finalize overridden on map's window

+++ app/src/main/java/com/cnet/VisualAnalysis/Fragments/VsmTransactionFragment.java
@@ -121,7 +121,7 @@ public class VsmTransactionFragment extends Fragment implements MainActivity.Key
 
         if (distributors > 0) {
             getSize();
-            handleDistDataChangeThread = new HandleDistributorChangeThread(changeDistributorHandler, distributors, (int) (vansSum(numberOfRowsInSingleVan.get(distributorIndex))), distributorStartIndex);
+            handleDistDataChangeThread = new HandleDistributorChangeThread(changeDistributorHandler, distributors, (vansSum(numberOfRowsInSingleVan.get(distributorIndex))), distributorStartIndex);
             handleDistDataChangeThread.start();
         }
     }
